Dev katlen

+++ src/main/java/com/accenture/bank/servico/AgenciaService.java
@@ -6,6 +6,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import com.accenture.bank.entity.Agencia;
+import com.accenture.bank.exceptions.AgenciaNaoEncontradaException;
+import com.accenture.bank.exceptions.ClienteNotFoundException;
+import com.accenture.bank.exceptions.FormatoIncorretoCliente;
+import com.accenture.bank.exceptions.FormatoInvalidoException;
 //import com.accenture.bank.entity.Agencia;
 import com.accenture.bank.repository.AgenciaRepository;
 //import com.accenture.bank.repository.ContaRepository;
@@ -21,6 +25,15 @@ public class AgenciaService {
 	AgenciaRepository contaCorrenteRepository;
 
 	public Agencia saveAgencia(Agencia agencia) {
+		
+		if( agencia.getNome().isBlank()||agencia.getNome().equals(null)||agencia.getNome().isEmpty()
+				||agencia.getTelefone().isBlank()||agencia.equals(null)||agencia.getTelefone().isEmpty()){
+			
+			throw new FormatoInvalidoException();
+			
+		}
+		
+				
 		// contaCorrenteRepository.saveAll(agencia.getContas());
 		return agenciaRepository.save(agencia);
 	}
@@ -32,6 +45,11 @@ public class AgenciaService {
 	}
 
 	public Agencia getById(Long id) {
+		
+		if(!agenciaRepository.existsById(id)) {
+			throw  new AgenciaNaoEncontradaException();
+		}
+		
 		return agenciaRepository.findById(id).get();
 	}
 
@@ -48,6 +66,10 @@ public class AgenciaService {
 	}
 
 	public void deletaAgencia(Long id) {
-		agenciaRepository.deleteById(id);
-	}
+		if(!agenciaRepository.existsById(id)) {
+			throw  new AgenciaNaoEncontradaException();
+		}
+			
+			agenciaRepository.deleteById(id);
+}
 }
